fix: drop expired events properly
If an event has expired and no one has interested in it anymore we’re
end up with an exception, which is accidentally caught by
event-assigner. It treats it as slave rejection, but in reality it is
not.

This led to the evil bug, where in the case of full queue there is no
way to flush it anymore, because each expired event will present in the
queue forever (until restarted).

This commit fixes that behaviour by dropping event properly.

Thanks @antmat for helping to find this bug!